Beginnings of Minix stdio.
I've imported the Minix stdio part of libc and I've got it to mostly
compile. There are a number of missing syscalls, e.g. lseek. So now
I might go back to the kernel and add as few extra syscalls as I can
to get libc working. The main aim is to get a decent editor (e.g. vi)
to work.