Major changes
I took a different approach for a bit. I extracted only the Linux syscall .o
files that matched the xv6 syscalls, and built a libc with these and the Minix
library code. This helped me to identify which functions were still missing
but which the Minix commands needed. I've written some dummy functions and
placed them in lib/xv.

I've merged this "linux testing" code back into the xv6-minix2 code, and
I've stared to bring over the simple Minix commands.

Right now, some of the Minix commands run, but most die after a bit. I am
guessing that there is a mismatch between the semantics of the xv6 kernel
syscalls and what the Minix code is expecting.